separate api tests and rename to test_ote_*
