+++ external/model-preparation-algorithm/tests/api_tests/test_ote_xai_api.py
@@ -5,16 +5,9 @@
 import os
 import os.path as osp
 
-import numpy as np
 import torch
 import pytest
 
-from mmcls.models import build_classifier
-from mmdet.models import build_detector
-
-from mpa.det.stage import DetectionStage  # noqa
-from mpa.modules.hooks.auxiliary_hooks import ReciproCAMHook, DetSaliencyMapHook
-from mpa.utils.config_utils import MPAConfig
 from mpa_tasks.apis.classification import ClassificationInferenceTask, ClassificationTrainTask
 from ote_sdk.entities.inference_parameters import InferenceParameters
 from ote_sdk.usecases.tasks.interfaces.export_interface import ExportType
@@ -22,9 +15,7 @@ from ote_sdk.entities.model import ModelEntity
 from ote_sdk.entities.train_parameters import TrainParameters
 from ote_sdk.entities.result_media import ResultMediaEntity
 from ote_sdk.test_suite.e2e_test_system import e2e_pytest_api
-from ote_sdk.test_suite.e2e_test_system import e2e_pytest_component
 from ote_cli.utils.io import save_model_data, read_model
-from ote_cli.registry import Registry
 from tests.api_tests.test_ote_classification_api import TestMPAClsAPI, DEFAULT_CLS_TEMPLATE_DIR
 from tests.api_tests.test_ote_detection_api import TestMPADetAPI, DEFAULT_DET_TEMPLATE_DIR
 from torchreid_tasks.openvino_task import OpenVINOClassificationTask
@@ -38,12 +29,6 @@ save_model_to = "/tmp/ote_xai/"
 
 assert_text_torch = "For the torch task the number of saliency maps should be equal to the number of all classes."
 assert_text_ov = "For the OV task the number of saliency maps should be equal to the number of predicted classes."
-
-templates_cls = Registry("external/model-preparation-algorithm").filter(task_type="CLASSIFICATION").templates
-templates_cls_ids = [template.model_template_id for template in templates_cls]
-
-templates_det = Registry("external/model-preparation-algorithm").filter(task_type="DETECTION").templates
-templates_det_ids = [template.model_template_id for template in templates_det]
 
 
 def saliency_maps_check(predicted_dataset, task_labels, assert_text, only_predicted=False):
@@ -62,84 +47,6 @@ def saliency_maps_check(predicted_dataset, task_labels, assert_text, only_predic
             assert saliency_map_counter == len(task_labels), assert_text
 
 
-class TestExplainMethods:
-    ref_saliency_vals_cls = {
-        "EfficientNet-B0": np.array([36, 185, 190, 159, 173, 124, 19], dtype=np.uint8),
-        "MobileNet-V3-large-1x": np.array([21, 38, 56, 134, 100, 41, 38], dtype=np.uint8),
-        "EfficientNet-V2-S": np.array([166, 204, 201, 206, 218, 221, 138], dtype=np.uint8),
-    }
-
-    ref_saliency_shapes = {
-        "ATSS": (2, 4, 4),
-        "SSD": (81, 13, 13),
-        "YOLOX": (80, 13, 13),
-    }
-
-    ref_saliency_vals_det = {
-        "ATSS": np.array([78, 217, 42, 102], dtype=np.uint8),
-        "SSD": np.array([225, 158, 221, 106, 146, 158, 227, 149, 137, 135, 200, 159, 255], dtype=np.uint8),
-        "YOLOX": np.array([109, 174, 82, 214, 178, 184, 168, 161, 163, 156, 220, 233, 195], dtype=np.uint8),
-    }
-
-    @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates_cls, ids=templates_cls_ids)
-    def test_saliency_map_cls(self, template):
-        base_dir = os.path.abspath(os.path.dirname(template.model_template_path))
-        cfg_path = os.path.join(base_dir, "model.py")
-        cfg = MPAConfig.fromfile(cfg_path)
-
-        cfg.model.pop("task")
-        model = build_classifier(cfg.model)
-        model = model.eval()
-
-        img = torch.rand(2, 3, 224, 224) - 0.5
-        data = {"img_metas": {}, "img": img}
-
-        with ReciproCAMHook(model) as rcam_hook:
-            with torch.no_grad():
-                _ = model(return_loss=False, **data)
-        saliency_maps = rcam_hook.records
-
-        assert len(saliency_maps) == 2
-        assert saliency_maps[0].ndim == 3
-        assert saliency_maps[0].shape == (1000, 7, 7)
-
-        assert (saliency_maps[0][0][0] == self.ref_saliency_vals_cls[template.name]).all()
-
-    @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates_det, ids=templates_det_ids)
-    def test_saliency_map_det(self, template):
-        base_dir = os.path.abspath(os.path.dirname(template.model_template_path))
-        cfg_path = os.path.join(base_dir, "model.py")
-        cfg = MPAConfig.fromfile(cfg_path)
-
-        model = build_detector(cfg.model)
-        model = model.eval()
-
-        img = torch.rand(2, 3, 416, 416) - 0.5
-        img_metas = [
-            {
-                "img_shape": (416, 416, 3),
-                "scale_factor": np.array([1.1784703, 0.832, 1.1784703, 0.832], dtype=np.float32),
-            },
-            {
-                "img_shape": (416, 416, 3),
-                "scale_factor": np.array([1.1784703, 0.832, 1.1784703, 0.832], dtype=np.float32),
-            },
-        ]
-        data = {"img_metas": [img_metas], "img": [img]}
-
-        with DetSaliencyMapHook(model) as det_hook:
-            with torch.no_grad():
-                _ = model(return_loss=False, rescale=True, **data)
-        saliency_maps = det_hook.records
-
-        assert len(saliency_maps) == 2
-        assert saliency_maps[0].ndim == 3
-        assert saliency_maps[0].shape == self.ref_saliency_shapes[template.name]
-        assert (saliency_maps[0][0][0] == self.ref_saliency_vals_det[template.name]).all()
-
-
 class TestOVClsXAIAPI(TestMPAClsAPI):
     @e2e_pytest_api
     @pytest.mark.parametrize(
@@ -149,8 +56,6 @@ class TestOVClsXAIAPI(TestMPAClsAPI):
     )
     def test_inference_xai(self, multilabel, hierarchical):
         hyper_parameters, model_template = self.setup_configurable_parameters(DEFAULT_CLS_TEMPLATE_DIR, num_iters=1)
-        multilabel = False
-        hierarchical = True
         task_environment, dataset = self.init_environment(
             hyper_parameters, model_template, multilabel, hierarchical, 20
         )
@@ -196,7 +101,6 @@ class TestOVClsXAIAPI(TestMPAClsAPI):
 class TestOVDetXAIAPI(TestMPADetAPI):
     @e2e_pytest_api
     def test_inference_xai(self):
-        save_model_to = "/tmp/ote_xai/"
         hyper_parameters, model_template = self.setup_configurable_parameters(DEFAULT_DET_TEMPLATE_DIR, num_iters=2)
         detection_environment, dataset = self.init_environment(hyper_parameters, model_template, 10)
